more refactoring, remove mainUI and socket stuff from about
