+++ polynote-frontend/polynote/util/functions.js
@@ -1,0 +1,20 @@
+'use strict';
+
+export function isEqual(a, b) {
+    if (a === b)
+        return true;
+
+    if (a instanceof Array || a instanceof Object) {
+        if (a.constructor !== b.constructor)
+            return false;
+
+        for (const i in a) {
+            if (!isEqual(a[i], b[i])) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    return false;
+}
